useEffect cleanup function example
Using a timer to check the form validity: previously, form validity was checked on every keystroke. This is overkill, particularly if (e.g.) it involved a HTTP request for validity or similar. By using a timer, and the subsequent cleanup return function to clear each timer, we can allow validation to happen only after a preset pause in input (in this case, 500ms). So, you can type in your email address and validity will be checked 500ms AFTER the typing stops, hence only running the code then (better!).